+++ LibraryManagement/src/main/java/com/boot/lms/LibraryManagement/controllerImp/BooksTransactionController_Imp.java
@@ -1,8 +1,10 @@
 package com.boot.lms.LibraryManagement.controllerImp;
 
+import com.boot.lms.LibraryManagement.Exception.UserNotFoundException;
 import com.boot.lms.LibraryManagement.Service.BooksTransactionService;
 import com.boot.lms.LibraryManagement.controller.BooksTransactionController;
 import com.boot.lms.LibraryManagement.objects.BooksTransaction_Request;
+import com.boot.lms.LibraryManagement.objects.UserRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -17,15 +19,24 @@ public class BooksTransactionController_Imp implements BooksTransactionControlle
     public List<BooksTransaction_Request> getAllTransactions() {
         return booksTransactionService.getAllTransactions();
     }
-
     @Override
     public BooksTransaction_Request getTransactionById(int id) {
-        return booksTransactionService.getTransactionById(id);
+        BooksTransaction_Request booksTransaction_request= booksTransactionService.getTransactionById(id);
+
+        if (booksTransaction_request == null)
+            throw new UserNotFoundException("BooksTransaction id not found-" + id);
+        return booksTransaction_request;
     }
 
     @Override
-    public void deleteTransaction(int id) {
+    public String deleteTransaction(int id) {
+
+
+        BooksTransaction_Request booksTransaction_request = booksTransactionService.getTransactionById(id);
+        if (booksTransaction_request == null)
+            throw new RuntimeException("BooksTransaction id not found" + id);
         booksTransactionService.deleteTransaction(id);
+        return "deleted BooksTransaction id- " + id;
 
     }
 }
